pyproject.toml
[tool.poetry]
name = "cap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.22.1"
pandas = "^1.3.5"
opencv-python = "^4.5.5"
nptyping = "^1.4.4"
pyserial = "^3.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
